Move activity components to plugin-activity
