+++ plugins/plugin-activity/che-plugin-activity-ide/src/main/resources/org/eclipse/che/plugin/activity/public/activity.js
@@ -1,0 +1,60 @@
+/*
+ * Copyright (c) 2012-2017 Codenvy, S.A.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *   Codenvy, S.A. - initial API and implementation
+ */
+var ActivityTracker = new function () {
+
+    var url;
+    var timeoutInterval = 30000;
+    var maxErrors = 10;
+    var errors = 0;
+    var active;
+
+    this.init = function (restContext, workspaceId) {
+        this.url = restContext + "/activity/" + workspaceId;
+        document.addEventListener("mousemove",  ActivityTracker.setActive);
+        document.addEventListener("keypress", ActivityTracker.setActive);
+        setInterval(ActivityTracker.sendRequest, timeoutInterval);
+
+    };
+
+    this.setActive = function() {
+        if (!active && errors < maxErrors) {
+            active = true;
+        }
+    }
+
+
+    this.sendRequest = function () {
+        if (!active) {
+            return;
+        }
+        active = false;
+
+        var request;
+        if (window.XMLHttpRequest) {
+            request = new XMLHttpRequest();
+        } else {
+            request = new ActiveXObject("Microsoft.XMLHTTP");
+        }
+
+        request.onreadystatechange = function () {
+            if (request.readyState == 4) {
+                if (request.status == 204) {
+                    errors = 0;
+                } else {
+                    errors++;
+                }
+
+            }
+        };
+        request.open("PUT", ActivityTracker.url, true);
+        request.send();
+    };
+};
