+++ plugins/plugin-activity/che-plugin-activity-server/src/main/test/org/eclipse/che/plugin/activity/WorkspaceActivityNotifierTest.java
@@ -1,0 +1,65 @@
+/*******************************************************************************
+ * Copyright (c) 2012-2017 Codenvy, S.A.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *   Codenvy, S.A. - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.che.plugin.activity;
+
+import org.eclipse.che.api.core.rest.HttpJsonRequestFactory;
+import org.mockito.Mock;
+import org.mockito.testng.MockitoTestNGListener;
+import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Listeners;
+import org.testng.annotations.Test;
+
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+
+/**
+ * Tests for {@link WorkspaceActivityNotifier}
+ *
+ * @author Mihail Kuznyetsov
+ */
+@Listeners(MockitoTestNGListener.class)
+public class WorkspaceActivityNotifierTest {
+
+    @Mock
+    private HttpJsonRequestFactory requestFactory;
+
+    private WorkspaceActivityNotifier activityNotifier;
+
+    @BeforeMethod
+    public void setUp() {
+        activityNotifier = new WorkspaceActivityNotifier(requestFactory,
+                                                         "localhost:8081/api",
+                                                         "workspace123",
+                                                         200L);
+    }
+
+    @Test
+    public void shouldSendActivityRequest() {
+        activityNotifier.onActivity();
+        verify(requestFactory).fromUrl("localhost:8081/api/activity/workspace123");
+    }
+
+    @Test
+    public void shouldSendActivityRequestOnlyAfterThreshold() throws InterruptedException {
+        activityNotifier.onActivity();
+        verify(requestFactory).fromUrl("localhost:8081/api/activity/workspace123");
+
+        Thread.sleep(50L);
+        activityNotifier.onActivity();
+
+        verify(requestFactory).fromUrl("localhost:8081/api/activity/workspace123");
+
+        Thread.sleep(200L);
+        activityNotifier.onActivity();
+
+        verify(requestFactory, times(2)).fromUrl("localhost:8081/api/activity/workspace123");
+    }
+}
