--- wsmaster/che-core-api-workspace/src/main/java/org/eclipse/che/api/workspace/server/activity/WorkspaceActivityService.java
@@ -1,68 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2012-2017 Codenvy, S.A.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   Codenvy, S.A. - initial API and implementation
- *******************************************************************************/
-package org.eclipse.che.api.workspace.server.activity;
-
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
-
-import org.eclipse.che.api.core.ForbiddenException;
-import org.eclipse.che.api.core.NotFoundException;
-import org.eclipse.che.api.core.ServerException;
-import org.eclipse.che.api.core.rest.Service;
-import org.eclipse.che.api.workspace.server.WorkspaceManager;
-import org.eclipse.che.api.workspace.server.model.impl.WorkspaceImpl;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.inject.Inject;
-import javax.inject.Singleton;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-
-import static org.eclipse.che.api.core.model.workspace.WorkspaceStatus.RUNNING;
-
-/**
- * Service for accessing API for updating activity timestamp of running workspaces.
- *
- * @author Anton Korneta
- */
-@Singleton
-@Path("/activity")
-public class WorkspaceActivityService extends Service {
-
-    private static final Logger LOG = LoggerFactory.getLogger(WorkspaceActivityService.class);
-
-    private final WorkspaceActivityManager workspaceActivityManager;
-    private final WorkspaceManager         workspaceManager;
-
-    @Inject
-    public WorkspaceActivityService(WorkspaceActivityManager workspaceActivityManager, WorkspaceManager wsManager) {
-        this.workspaceActivityManager = workspaceActivityManager;
-        this.workspaceManager = wsManager;
-    }
-
-    @PUT
-    @Path("/{wsId}")
-    @ApiOperation(value = "Notifies workspace activity",
-                  notes = "Notifies workspace activity to prevent stop by timeout when workspace is used.")
-    @ApiResponses(@ApiResponse(code = 204, message = "Activity counted"))
-    public void active(@ApiParam(value = "Workspace id")
-                       @PathParam("wsId") String wsId) throws ForbiddenException, NotFoundException, ServerException {
-        final WorkspaceImpl workspace = workspaceManager.getWorkspace(wsId);
-        if (workspace.getStatus() == RUNNING) {
-            workspaceActivityManager.update(wsId, System.currentTimeMillis());
-            LOG.debug("Updated activity on workspace {}", wsId);
-        }
-    }
-}
